Delete commentaire implémenté

+++ YoutubeClone/Controllers/CommentairesController.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using System.Net;
 using System.Web;
 using System.Web.Mvc;
+using System.Web.Routing;
 using YoutubeClone.Models.Data_Models;
 
 namespace YoutubeClone.Controllers
@@ -13,28 +14,6 @@ namespace YoutubeClone.Controllers
     public class CommentairesController : Controller
     {
         private YoutubeCloneDb db = new YoutubeCloneDb();
-
-        // GET: Commentaires
-        public ActionResult Index()
-        {
-            var commentaires = db.Commentaires.Include(c => c.Auteur).Include(c => c.Video);
-            return View(commentaires.ToList());
-        }
-
-        // GET: Commentaires/Details/5
-        public ActionResult Details(string id)
-        {
-            if (id == null)
-            {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            Commentaire commentaire = db.Commentaires.Find(id);
-            if (commentaire == null)
-            {
-                return HttpNotFound();
-            }
-            return View(commentaire);
-        }
 
         // GET: Commentaires/Create
         public ActionResult Create()
@@ -77,66 +56,31 @@ namespace YoutubeClone.Controllers
             return View(commentaire);
         }
 
-        // GET: Commentaires/Edit/5
-        public ActionResult Edit(string id)
-        {
-            if (id == null)
-            {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            Commentaire commentaire = db.Commentaires.Find(id);
-            if (commentaire == null)
-            {
-                return HttpNotFound();
-            }
-            ViewBag.Chaine_FK = new SelectList(db.Chaines, "ChaineId", "Name", commentaire.Utilisateur_FK);
-            ViewBag.Video_FK = new SelectList(db.Videos, "VideoId", "Name", commentaire.Video_FK);
-            return View(commentaire);
-        }
-
-        // POST: Commentaires/Edit/5
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "CommentaireId,Contenu,DatePublication,Chaine_FK,Video_FK")] Commentaire commentaire)
-        {
-            if (ModelState.IsValid)
-            {
-                db.Entry(commentaire).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
-            }
-            ViewBag.Chaine_FK = new SelectList(db.Chaines, "ChaineId", "Name", commentaire.Utilisateur_FK);
-            ViewBag.Video_FK = new SelectList(db.Videos, "VideoId", "Name", commentaire.Video_FK);
-            return View(commentaire);
-        }
-
         // GET: Commentaires/Delete/5
-        public ActionResult Delete(string id)
+        [Authorize]
+        public ActionResult Delete(int id)
         {
-            if (id == null)
-            {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
             Commentaire commentaire = db.Commentaires.Find(id);
-            if (commentaire == null)
+            Video v = db.Videos.Find(commentaire.Video_FK);
+            Utilisateur u = db.Utilisateurs.FirstOrDefault(user => user.Username == User.Identity.Name);
+            if (u.Username == v.Channel.Utilisateur.Username || u.IsAdmin)
             {
-                return HttpNotFound();
+                db.Commentaires.Remove(commentaire);
+                db.SaveChanges();
             }
-            return View(commentaire);
+            return RedirectToAction("Details", "Videos", new { id = commentaire.Video_FK });  
         }
 
         // POST: Commentaires/Delete/5
-        [HttpPost, ActionName("Delete")]
-        [ValidateAntiForgeryToken]
-        public ActionResult DeleteConfirmed(string id)
-        {
-            Commentaire commentaire = db.Commentaires.Find(id);
-            db.Commentaires.Remove(commentaire);
-            db.SaveChanges();
-            return RedirectToAction("Index");
-        }
+        //[HttpPost, ActionName("Delete")]
+        //[ValidateAntiForgeryToken]
+        //public ActionResult DeleteConfirmed(int id)
+        //{
+        //    Commentaire commentaire = db.Commentaires.Find(id);
+        //    db.Commentaires.Remove(commentaire);
+        //    db.SaveChanges();
+        //    return RedirectToAction("Index");
+        //}
 
         protected override void Dispose(bool disposing)
         {
